+++ conf/script/build_system/compiler/__init__.py
@@ -1,2 +1,2 @@
-from build_system.compiler.compiler_family import CompilerFamily as Family
+from build_system.compiler.family import CompilerFamily as Family
 from build_system.compiler.version import CompilerVersion as Version
