+++ conf/script/build_system/compiler/host/get_info/cli.py
@@ -4,7 +4,7 @@ from typing import Any, AnyStr, Callable, Final, Optional
 
 import colorama
 
-from build_system.compiler.compiler_family import CompilerFamily
+from build_system.compiler.family import CompilerFamily
 
 
 def cli_init():
