+++ conf/script/build_system/compiler/host/get_info/version/clang/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import build_system.compiler.host.get_info.version.gnu.cli
-from build_system.compiler.compiler_family import CompilerFamily
+from build_system.compiler.family import CompilerFamily
 from build_system.compiler.host.get_info import cli, version
 
 # Run as a script
