+++ conf/script/build_system/compiler/host/get_info/version/gcc/__main__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from build_system.compiler.compiler_family import CompilerFamily
+from build_system.compiler.family import CompilerFamily
 from build_system.compiler.host.get_info import cli, version
 from build_system.compiler.host.get_info.version import gnu
